requirements.txt
amqp==5.3.1
anyio==4.6.0
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
beautifulsoup4==4.12.3
billiard==4.2.1
bs4==0.0.2
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cron-descriptor==1.4.5
cryptography==44.0.0
daphne==4.1.2
Django==5.1
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-filter==24.3
django-sslserver==0.22
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dnspython==2.7.0
eventlet==0.38.0
exceptiongroup==1.2.2
greenlet==3.1.1
h11==0.14.0
holehe==1.61
httpcore==1.0.5
httpx==0.27.2
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
kombu==5.4.2
outcome==1.3.0.post0
pillow==10.4.0
prompt_toolkit==3.0.48
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==24.3.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-http-client==3.3.7
pytz==2024.2
redis==5.2.1
requests==2.32.3
sendgrid==6.11.0
service-identity==24.2.0
setuptools==75.6.0
sib-api-v3-sdk==7.6.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.5.1
starkbank-ecdsa==2.2.0
stripe==11.2.0
termcolor==2.4.0
tomli==2.2.1
tqdm==4.66.5
trio==0.26.2
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
zope.interface==7.2
